feat(dde-control-center): 删除无用代码，修改代码中出现的警告,泄露，空指针问题
删除一些无用代码，以及一些由于变量为初始化，赋值类型不一致导致的警告,泄露，空指针问题问题

Log: 控制中心废弃代码删除,代码出现警告,泄露，空指针问题

Bug: https://pms.uniontech.com/zentao/task-view-27932.html

+++ src/frame/modules/display/monitorindicator.cpp
@@ -50,16 +50,16 @@ void MonitorIndicator::resizeEvent(QResizeEvent *e)
     XRectangle rectangle;
     rectangle.x = 0;
     rectangle.y = 0;
-    rectangle.width = e->size().width();
-    rectangle.height = e->size().height();
+    rectangle.width = static_cast<ushort>(e->size().width());
+    rectangle.height = static_cast<ushort>(e->size().height());
 
     // need to restore the cut area, if not,cut out will be repeated.
     XShapeCombineRectangles(QX11Info::display(), winId(), ShapeBounding, 0, 0, &rectangle, 1, ShapeSet, YXBanded);
 
     rectangle.x = 10;
     rectangle.y = 10;
-    rectangle.width = e->size().width() - 20;
-    rectangle.height = e->size().height() - 20;
+    rectangle.width = static_cast<ushort>(e->size().width()) - 20;
+    rectangle.height = static_cast<ushort>(e->size().height()) - 20;
 
     XShapeCombineRectangles(QX11Info::display(), winId(), ShapeBounding, 0, 0, &rectangle, 1, ShapeSubtract, YXBanded);
 }
